break codes in visual display into smaller files
