+++ src/Covid19/ShowMenu.java
@@ -53,7 +53,7 @@ public class ShowMenu {
         }
     }
 
-    // Users can continue as many times as they want.
+    // Method that let users can continue the program as many times as they want.
     // OR: They can choose to end the program when they finish.
     public static void FinalChoice() throws ParseException, IOException {
         while (true) {
@@ -62,7 +62,8 @@ public class ShowMenu {
             int DataChoice = getDataChoice();
             runIndividualProgram(DataChoice);
             //Continue or stop the program - Final Choice
-            System.out.println("\n\nThis is the end of the program. Do you want to restart or not? ");
+            System.out.println("\n\n_____________________________________________________________________________________");
+            System.out.println("\nThis is the end of the program. Do you want to restart or not? ");
             System.out.println("1. Restart the program");
             System.out.println("2. End the program");
             do {
@@ -81,4 +82,5 @@ public class ShowMenu {
             }
         }
     }
+
 }
